+++ proxy/internal/cookiemanager/cookiemanager_test.go
@@ -77,7 +77,7 @@ func TestCookies(t *testing.T) {
 	}
 
 	recorder = httptest.NewRecorder()
-	if err := cm.SetIDTokenCookie(recorder, req, tok); err != nil {
+	if err := cm.SetIDTokenCookie(recorder, req, tok, nil); err != nil {
 		t.Fatal("cookie not set")
 	}
 	v = recorder.Header().Get("Set-Cookie")